# swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "version": "1.0.0",
        "description": "API for Social-Media Application",
        "title": "Social-Media API"
    },
    "servers": [
        {
            "url": "http://localhost:3100"
        }
    ],
    "components": {
        "securitySchemes": {
            "JWT": {
                "type": "apiKey",
                "in": "header",
                "name": "Authorization"
            }
        }
    },
    "paths": {
...
            "/api/posts/": {
              "get": {
                "tags": ["Posts"],
                "summary": "Get posts by user ID",
                "description": "Retrieve posts by user ID",
                "security": [
                  {
                    "JWT": []
                  }
                ],
                
                "responses": {
                  "200": {
                    "description": "OK"
                  },
                  "404": {
                    "description": "No posts found for the user"
                  }
                }
              },
              "post": {
                "tags": ["Posts"],
                "summary": "Add a new post",
                "description": "Add a new post with an image",
                "security": [
                  {
                    "JWT": []
                  }
                ],
                "requestBody": {
                  "content": {
                    "multipart/form-data": {
                      "schema": {
                        "type": "object",
                        "properties": {
                          "imgUrl": {
                            "type": "string",
                            "format": "binary"
                          },
                          "caption": {
                            "type": "string"
                          }
                          
                        },
                        "required": ["imgUrl", "caption"]
                      }
                    }
                  }
                },
                "responses": {
                  "200": {
                    "description": "OK"
                  },
                  "400": {
                    "description": "Bad request"
                  }
                }
              }
            },
...
            "/api/comments/": {
              "post": {
                "tags": ["Comments"],
                "summary": "Add a new comment",
                "description": "Add a new comment to a post",
                "security": [
                  {
                    "JWT": []
                  }
                ],
                "requestBody": {
                  "content": {
                    "application/json": {
                      "schema": {
                        "type": "object",
                        "properties": {
                          "postId": {
                            "type": "integer"
                          },
                          "text": {
                            "type": "string"
                          }
                        },
                        "required": ["postId", "text"]
                      }
                    }
                  }
                },
                "responses": {
                  "200": {
                    "description": "OK"
                  },
                  "400": {
                    "description": "Bad request"
                  }
                }
              },
              "put": {
                "tags": ["Comments"],
                "summary": "Update a comment",
                "description": "Update an existing comment",
                "security": [
                  {
                    "JWT": []
                  }
                ],
                "requestBody": {
                  "content": {
                    "application/json": {
                      "schema": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "integer"
                          },
                          "text": {
                            "type": "string"
                          }
                        },
                        "required": ["id", "text"]
                      }
                    }
                  }
                },
                "responses": {
                  "200": {
                    "description": "OK"
                  },
                  "404": {
                    "description": "Comment not found"
                  }
                }
              }
            },
...
    }
}